Refactor and improve code quality across multiple modules
- Introduced `safe_positive_int` utility function to ensure safe conversion of values to positive integers in `pcbs_validation.py` and `search_benchmark_utils.py`.
- Updated various calculations in `pcbs_validation.py` to use `safe_positive_int` for better error handling.
- Enhanced readability by replacing special characters with standard symbols (e.g., "->" instead of "â†’") in comments and docstrings across several files, including `evaluator.py`, `generator.py`, and `cognitive_bounded_search.py`.
- Improved logging messages in `solver_comparison.py` to use clearer status indicators.
- Added context management for `torch.no_grad()` in `repair_feedback.py` to optimize memory usage during inference.
- Updated the Zelda data loader to use a default set of training dungeons instead of hardcoded values.
- Fixed minor issues in test files to ensure consistency with updated comments and patterns.

+++ src/core/vqvae.py
@@ -1113,7 +1113,7 @@ class SemanticVQVAE(nn.Module):
         """
         Differentiable soft penalty over 3x3 neighborhood illegal adjacencies.
         """
-        probs = F.softmax(recon_logits, dim=1)
+        probs = F.softmax(recon_logits.float(), dim=1).to(dtype=recon_logits.dtype)
         illegal = self.illegal_adjacency_matrix.to(dtype=probs.dtype, device=probs.device)
 
         # 8-neighborhood shifts (3x3 window excluding center).
@@ -1473,7 +1473,7 @@ class SemanticVQVAE2(nn.Module):
         return torch.zeros((batch_size, height, width), dtype=torch.long, device=device)
 
     def _illegal_adjacency_penalty(self, recon_logits: Tensor) -> Tensor:
-        probs = F.softmax(recon_logits, dim=1)
+        probs = F.softmax(recon_logits.float(), dim=1).to(dtype=recon_logits.dtype)
         illegal = self.illegal_adjacency_matrix.to(dtype=probs.dtype, device=probs.device)
         shifts = [
             (-1, -1), (-1, 0), (-1, 1),
